pyproject.toml
[tool.poetry]
name = "x-vault_server"
version = "0.1.0"
description = ""
authors = ["nl <nicnac9911@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyNaCl = "^1.4.0"
passlib = "^1.7.4"
argon2-cffi = "^21.1.0"
PyYAML = "^5.4.1"
Flask = "^2.0.2"
Flask-Tortoise = "^0.0.2"
aerich = "^0.5.8"
Flask-Session = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
uvicorn = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
